requirements.txt
amqp==5.0.2
apipkg==1.5
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
attrs==20.3.0
autopep8==1.5.4
awscli==1.18.217
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
black==20.8b1
blinker==1.4
boto3==1.16.56
botocore==1.19.63
Brlapi==0.7.0
brunette==0.2.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.0
celery==5.0.4
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
contextlib2==0.6.0
coverage==5.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-channels==0.7.0
django-ckeditor==6.0.0
django-compat==1.0.15
django-cors-headers==3.6.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-localflavor==3.0.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-registration==3.1.1
django-sendgrid-v5==0.9.0
django-settings-export==1.2.1
django-storages==1.10.1
django-tailwind==1.1.0
django-timezone-field==4.1.1
django-tinymce==3.0.2
djangorestframework==3.11.1
docker==4.4.1
docker-compose==1.28.0rc1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
execnet==1.8.0
fastapi==0.63.0
fasteners==0.14.1
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
future==0.18.2
h11==0.12.0
html5lib==1.0.1
httplib2==0.14.0
httptools==0.1.1
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
ipaddr==2.2.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
iso-639==0.4.5
iso8601==0.1.13
jedi==0.18.0
jmespath==0.9.4
jsonschema==3.2.0
kazam==1.4.5
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
m3u8==0.7.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
mypy-extensions==0.4.3
mysqlclient==2.0.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
parso==0.8.1
pathspec==0.8.1
pep517==0.8.2
pep8-naming==0.11.1
pexpect==4.6.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.7
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pydantic==1.7.3
pyflakes==2.2.0
Pygments==2.3.1
PyGObject==3.36.0
pygraphviz==1.5
PyJWT==1.7.1
pymacaroons==0.13.0
pymemcache==3.4.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.17.3
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-mock==3.3.1
pytest-voluptuous==1.2.0
pytest-xdist==2.1.0
python-apt==2.0.0+ubuntu0.20.4.3
python-crontab==2.5.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
python-http-client==3.3.1
python-stdnum==1.14
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
razorpay==1.2.0
redis==3.5.3
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
roman==2.0.0
rope==0.18.0
rsa==4.0
s3transfer==0.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sendgrid==6.5.0
sentry-sdk==0.19.5
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
starlette==0.13.6
systemd-python==234
texttable==1.6.3
toml==0.10.1
traitlets==5.0.5
ttystatus==0.38
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.13.3
uvloop==0.14.0
uWSGI==2.0.19.1
vine==5.0.0
virtualenv==20.0.17
voluptuous==0.12.1
wadllib==1.3.3
watchgod==0.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
whitenoise==5.2.0
xkit==0.0.0
youtube-dl==2021.1.3
zipp==1.0.0
